requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==6.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
fastapi==0.124.0
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
outcome==1.3.0.post0
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.5
selenium==4.39.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.4
starlette==0.50.0
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
websocket-client==1.9.0
wsproto==1.3.2
